refactor: re-arrange structure of the store and typings
