+++ src/store/interfaces/IContent.ts
