--- src/typings/index.d.ts
@@ -1,1 +1,0 @@
-declare module "react-uuid";
